backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
sendgrid==6.7.1
twilio==6.60.0
requests==2.25.1
openai==1.3.5
python-datauri==2.2.0
python-dotenv==0.19.1
Pillow==9.5.0
Werkzeug==2.2.3
SQLAlchemy==1.4.46